+++ edgeworkers/testdata/edgekv_with_staging_network_no_group_id/edgekv.tf
@@ -1,0 +1,21 @@
+terraform {
+  required_providers {
+    akamai = {
+      source = "akamai/akamai"
+    }
+  }
+  required_version = ">= 0.13"
+}
+
+provider "akamai" {
+  edgerc         = var.edgerc_path
+  config_section = var.config_section
+}
+
+resource "akamai_edgekv" "edgekv" {
+  namespace_name       = "test_namespace"
+  network              = "staging"
+  group_id             = 0
+  retention_in_seconds = 0
+  geo_location         = "EU"
+}