+++ edgeworkers/testdata/edgeworker_with_no_local_bundle/variables.tf
@@ -1,0 +1,14 @@
+variable "edgerc_path" {
+  type    = string
+  default = "~/.edgerc"
+}
+
+variable "config_section" {
+  type    = string
+  default = "test_section"
+}
+
+variable "env" {
+  type    = string
+  default = "STAGING"
+}